+++ player.py
@@ -13,7 +13,7 @@ class Player:
         self.rel = 0
         self.health_recovery_delay = 1
         self.time_prev = pg.time.get_ticks()
-        self.armor = 0
+        self.armor = 50
 
     def recovery_health(self):
         if self.check_health_recovry_delay() and self.health < PLAYER_MAX_HEALTH:
